前后端修改，并增加了上传服装功能

+++ backend/recommender/StatisticsRecommender/src/main/scala/com/tqz/statistics/StatisticsRecommender.scala
@@ -45,15 +45,15 @@ object StatisticsRecommender {
     ratingDF.createOrReplaceTempView("ratings")
 
     // 用spark sql去做不同的统计推荐
-    // 1. 历史热门商品，按照评分个数统计，productId，count
+    // 1. 历史热门商品，按照评分个数统计，productId，count(推荐在大家都在买)
     val rateMoreProductsDF = spark.sql("select productId, count(productId) as count from ratings group by productId order by count desc")
     storeDFInMongoDB(rateMoreProductsDF, RATE_MORE_PRODUCTS)
 
-    // 2. 近期热门商品，把时间戳转换成yyyyMM格式进行评分个数统计，最终得到productId, count, yearmonth
+    // 2. 近期热门商品，把时间戳转换成yyyyMM格式进行评分个数统计，最终得到productId, count, yearmonth(推荐在大家都在看)
     // 创建一个日期格式化工具
     val simpleDateFormat = new SimpleDateFormat("yyyyMM")
     // 注册UDF，将timestamp转化为年月格式yyyyMM
-    spark.udf.register("changeDate", (x: Int) => simpleDateFormat.format(new Date(x * 1000L)).toInt)
+    spark.udf.register("changeDate", (x: Long) => simpleDateFormat.format(new Date(x * 1000L)).toLong)
     // 把原始rating数据转换成想要的结构productId, score, yearmonth
     val ratingOfYearMonthDF = spark.sql("select productId, score, changeDate(timestamp) as yearmonth from ratings")
     ratingOfYearMonthDF.createOrReplaceTempView("ratingOfMonth")
